[spark] Add test for InMemoryColumnStore and refactor common test code
